Added authentication for API

+++ routes/api.php
@@ -1,15 +1,14 @@
 <?php
 
-use Illuminate\Http\Request;
+use App\Http\Controllers\Api\AuthController;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\FixtureController;
 
-//Route::post('/tokens/create', function (Request $request) {
-//    return $request->user()->createToken('token-name', ['server:update'])->plainTextToken;
-//})->middleware('auth:sanctum');
+Route::post('/auth/register', [AuthController::class, 'createUser']);
+Route::post('/login', [AuthController::class, 'loginUser']);
 
-Route::get('/fixtures', [FixtureController::class, 'index']);
-Route::get('/fixtures/{id}', [FixtureController::class, 'show']);
-Route::delete('/fixtures/{id}', [FixtureController::class, 'delete']);
-Route::post('/fixtures', [FixtureController::class, 'save']);
-Route::post('/json-escaper', [FixtureController::class, 'escape']);
+Route::get('/fixtures', [FixtureController::class, 'index'])->middleware(['auth:sanctum', 'ability:read-api,write-api']);
+Route::get('/fixtures/{id}', [FixtureController::class, 'show'])->middleware(['auth:sanctum', 'ability:read-api,write-api']);
+Route::delete('/fixtures/{id}', [FixtureController::class, 'delete'])->middleware(['auth:sanctum', 'abilities:write-api']);
+Route::post('/fixtures', [FixtureController::class, 'save'])->middleware(['auth:sanctum', 'abilities:write-api']);
+Route::post('/json-escaper', [FixtureController::class, 'escape'])->middleware(['auth:sanctum', 'abilities:write-api']);
